+++ lib/post/post_detail/view/widget/post_detail.dart
@@ -2,12 +2,12 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:together_delivery_app/post_detail/view/widget/post_detail_content.dart';
-import 'package:together_delivery_app/post_detail/view/widget/post_detail_extra_info.dart';
-import 'package:together_delivery_app/post_detail/view/widget/post_detail_image.dart';
-import 'package:together_delivery_app/post_detail/view/widget/post_detail_meet_location.dart';
-import 'package:together_delivery_app/post_detail/view/widget/post_detail_restaurantName.dart';
-import 'package:together_delivery_app/post_detail/view/widget/post_header.dart';
+import 'package:together_delivery_app/post/post_detail/view/widget/post_detail_content.dart';
+import 'package:together_delivery_app/post/post_detail/view/widget/post_detail_extra_info.dart';
+import 'package:together_delivery_app/post/post_detail/view/widget/post_detail_image.dart';
+import 'package:together_delivery_app/post/post_detail/view/widget/post_detail_meet_location.dart';
+import 'package:together_delivery_app/post/post_detail/view/widget/post_detail_restaurantName.dart';
+import 'package:together_delivery_app/post/post_detail/view/widget/post_header.dart';
 
 import '../../model/post_detail_response_model.dart';
