+++ lib/post/post_detail/view/widget/post_detail_content.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:together_delivery_app/post_detail/model/post_detail_response_model.dart';
+
+import '../../model/post_detail_response_model.dart';
 
 class PostContent extends ConsumerWidget {
   final Post post;
