+++ lib/post/post_detail/view/widget/post_detail_extra_info.dart
@@ -3,7 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:together_delivery_app/common/const/restaurantCategory.dart';
 import 'package:together_delivery_app/common/util/money_convertor.dart';
-import 'package:together_delivery_app/post_detail/model/post_detail_response_model.dart';
+
+import '../../model/post_detail_response_model.dart';
 
 class PostDetailExtraInfo extends ConsumerWidget {
   final Post post;
