+++ lib/post/post_detail/view/widget/post_detail_meet_location.dart
@@ -4,7 +4,9 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_naver_map/flutter_naver_map.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:together_delivery_app/post_detail/model/post_detail_response_model.dart';
+
+import '../../model/post_detail_response_model.dart';
+
 
 class PostDetailMeetLocation extends ConsumerWidget {
   final MeetLocation meetLocation;
