+++ lib/post/post_detail/view/widget/post_header.dart
@@ -2,9 +2,9 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:together_delivery_app/common/util/dataConvertor.dart';
-import 'package:together_delivery_app/post_detail/model/post_detail_response_model.dart';
 
 import '../../../post_list/view/widget/post_status.dart';
+import '../../model/post_detail_response_model.dart';
 
 class PostDetailHeader extends ConsumerWidget {
   final Post post;
