+++ lib/post/post_list/view/post_list_view.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
-import 'package:together_delivery_app/post_list/model/api/post_list_response.dart';
-import 'package:together_delivery_app/post_list/view/widget/post_list_error_indicator.dart';
-import 'package:together_delivery_app/post_list/view/widget/post_list_item.dart';
-import 'package:together_delivery_app/post_list/view/widget/no_post_item_indicator.dart';
+import 'package:together_delivery_app/post/post_list/view/widget/no_post_item_indicator.dart';
+import 'package:together_delivery_app/post/post_list/view/widget/post_list_error_indicator.dart';
+import 'package:together_delivery_app/post/post_list/view/widget/post_list_item.dart';
 
+import '../model/api/post_list_response.dart';
 import '../view_model/post_list_view_model.dart';
 
 class PostListView extends ConsumerStatefulWidget {
@@ -31,9 +31,9 @@ class _PostListViewState extends ConsumerState<PostListView> {
       postListViewModel.getPostList(pageKey).then((response) {
         final isLastPage = response.metaData.currentPage >= response.metaData.totalPage;
         if (isLastPage) {
-          _pagingController.appendLastPage(response.posts);
+          _pagingController.appendLastPage(response.posts.cast<Post>());
         } else {
-          _pagingController.appendPage(response.posts, pageKey + 1);
+          _pagingController.appendPage(response.posts.cast<Post>(), pageKey + 1);
         }
       }).catchError((error) {
         _pagingController.error = error;
