+++ lib/post/post_list/view/widget/post_item.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/cupertino.dart';
-import 'package:together_delivery_app/common/routes/routes.dart';
-import 'package:together_delivery_app/post_list/model/api/post_list_response.dart';
-import 'package:together_delivery_app/post_list/view/widget/post_item_content.dart';
-import 'package:together_delivery_app/post_list/view/widget/post_item_header.dart';
-import 'package:together_delivery_app/post_list/view/widget/post_item_restaurant_info.dart';
+import 'package:together_delivery_app/post/post_list/view/widget/post_item_content.dart';
+import 'package:together_delivery_app/post/post_list/view/widget/post_item_header.dart';
+import 'package:together_delivery_app/post/post_list/view/widget/post_item_restaurant_info.dart';
+
+import '../../model/api/post_list_response.dart';
 
 class PostItem extends StatelessWidget {
   final Post post;
