+++ lib/post/post_list/view/widget/post_item_restaurant_info.dart
@@ -3,7 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:together_delivery_app/common/config/color/app_color.dart';
 import 'package:together_delivery_app/common/const/restaurantCategory.dart';
 import 'package:together_delivery_app/common/util/money_convertor.dart';
-import 'package:together_delivery_app/post_list/model/api/post_list_response.dart';
+
+import '../../model/api/post_list_response.dart';
+
 
 class PostItemRestaurantInfo extends StatelessWidget {
   final Post post;
