+++ lib/post/post_list/view/widget/post_list_item.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/cupertino.dart';
-import 'package:together_delivery_app/post_list/model/api/post_list_response.dart';
-import 'package:together_delivery_app/post_list/view/widget/post_item.dart';
-import 'package:together_delivery_app/post_list/view/widget/post_list_Item_wrapper.dart';
+import 'package:together_delivery_app/post/post_list/view/widget/post_item.dart';
+import 'package:together_delivery_app/post/post_list/view/widget/post_list_Item_wrapper.dart';
+
+import '../../model/api/post_list_response.dart';
 
 class PostListItem extends StatelessWidget {
   final Post post;
