+++ lib/post/repository/post_repository.dart
@@ -1,6 +1,5 @@
-import 'package:together_delivery_app/post_list/model/api/post_list_response.dart';
-
-import '../../post_detail/model/post_detail_response_model.dart';
+import '../post_detail/model/post_detail_response_model.dart';
+import '../post_list/model/api/post_list_response.dart';
 
 abstract class PostRepository {
   Future<PostListResponse> getPostList(String page);
